move ::stream into ::net::stream
net will contain all network modules agnostic of transport,
including also a user module for stuff like credentials,
and userId, for start
