+++ examples/https_connect_proxy.rs
@@ -35,9 +35,9 @@ use rama::{
         server::HttpServer,
         Body, IntoResponse, Request, RequestContext, Response, StatusCode,
     },
+    net::stream::layer::http::BodyLimitLayer,
     rt::Executor,
     service::{service_fn, Context, Service, ServiceBuilder},
-    stream::layer::http::BodyLimitLayer,
     tcp::{server::TcpListener, utils::is_connection_error},
     tls::{
         dep::rcgen::KeyPair,
