+++ rama-cli/src/cmd/echo/mod.rs
@@ -10,13 +10,13 @@ use rama::{
         server::HttpServer,
         IntoResponse, Request, RequestContext, Response,
     },
+    net::stream::{layer::http::BodyLimitLayer, SocketInfo},
     proxy::pp::server::HaProxyLayer,
     rt::Executor,
     service::{
         layer::{limit::policy::ConcurrentPolicy, LimitLayer, TimeoutLayer},
         Context, ServiceBuilder,
     },
-    stream::{layer::http::BodyLimitLayer, SocketInfo},
     tcp::server::TcpListener,
     tls::rustls::server::IncomingClientHello,
     ua::{UserAgent, UserAgentClassifierLayer},
