+++ rama-cli/src/cmd/proxy/mod.rs
@@ -14,12 +14,12 @@ use rama::{
         server::HttpServer,
         Body, IntoResponse, Request, RequestContext, Response, StatusCode,
     },
+    net::stream::layer::http::BodyLimitLayer,
     rt::Executor,
     service::{
         layer::{limit::policy::ConcurrentPolicy, LimitLayer, TimeoutLayer},
         service_fn, Context, Service, ServiceBuilder,
     },
-    stream::layer::http::BodyLimitLayer,
     tcp::{server::TcpListener, utils::is_connection_error},
 };
 use std::{convert::Infallible, time::Duration};
