+++ src/http/layer/opentelemetry.rs
@@ -13,8 +13,8 @@ use crate::{
         headers::{HeaderMapExt, UserAgent},
         IntoResponse, Request, RequestContext, Response,
     },
+    net::stream::SocketInfo,
     service::{Context, Layer, Service},
-    stream::SocketInfo,
 };
 use headers::ContentLength;
 use std::{fmt, sync::Arc, time::SystemTime};
@@ -34,7 +34,7 @@ const HTTP_SERVER_ACTIVE_REQUESTS: &str = "http.server.active_requests";
 // TODO: do we also want to track actual calculated body size?
 // this would mean we _need_ to buffer the body, which is not ideal
 // Perhaps make it opt-in?
-// NOTE: we could also make this opt-in via BytesRWTrackerHandle (rama::stream::BytesRWTrackerHandle)
+// NOTE: we could also make this opt-in via BytesRWTrackerHandle (rama::net::stream::BytesRWTrackerHandle)
 // this would however not work properly (I think) with h2/h3...
 // const HTTP_SERVER_REQUEST_SIZE: &str = "http.server.request.size";
 // const HTTP_SERVER_RESPONSE_SIZE: &str = "http.server.response.size";
