+++ src/net/mod.rs
@@ -1,0 +1,3 @@
+//! protocol agnostic network modules
+
+pub mod stream;
