+++ src/net/stream/matcher/ip.rs
@@ -1,9 +1,9 @@
 use http::Request;
 
-use crate::stream::dep::ipnet::{IpNet, Ipv4Net, Ipv6Net};
+use crate::net::stream::dep::ipnet::{IpNet, Ipv4Net, Ipv6Net};
 use crate::{
+    net::stream::SocketInfo,
     service::{context::Extensions, Context},
-    stream::SocketInfo,
 };
 
 #[derive(Debug, Clone)]
@@ -56,7 +56,7 @@ impl<State, Body> crate::service::Matcher<State, Request<Body>> for IpNetMatcher
 
 impl<State, Socket> crate::service::Matcher<State, Socket> for IpNetMatcher
 where
-    Socket: crate::stream::Socket,
+    Socket: crate::net::stream::Socket,
 {
     fn matches(
         &self,
@@ -265,7 +265,7 @@ mod test {
             peer_addr: Option<SocketAddr>,
         }
 
-        impl crate::stream::Socket for FakeSocket {
+        impl crate::net::stream::Socket for FakeSocket {
             fn local_addr(&self) -> std::io::Result<SocketAddr> {
                 match &self.local_addr {
                     Some(addr) => Ok(*addr),
