+++ src/net/stream/matcher/loopback.rs
@@ -1,8 +1,8 @@
 use http::Request;
 
 use crate::{
+    net::stream::SocketInfo,
     service::{context::Extensions, Context},
-    stream::SocketInfo,
 };
 
 #[derive(Debug, Clone)]
@@ -61,7 +61,7 @@ impl<State, Body> crate::service::Matcher<State, Request<Body>> for LoopbackMatc
 
 impl<State, Socket> crate::service::Matcher<State, Socket> for LoopbackMatcher
 where
-    Socket: crate::stream::Socket,
+    Socket: crate::net::stream::Socket,
 {
     fn matches(
         &self,
@@ -140,7 +140,7 @@ mod test {
             peer_addr: Option<SocketAddr>,
         }
 
-        impl crate::stream::Socket for FakeSocket {
+        impl crate::net::stream::Socket for FakeSocket {
             fn local_addr(&self) -> std::io::Result<SocketAddr> {
                 match &self.local_addr {
                     Some(addr) => Ok(*addr),
