+++ src/net/stream/matcher/port.rs
@@ -1,8 +1,8 @@
 use http::Request;
 
 use crate::{
+    net::stream::SocketInfo,
     service::{context::Extensions, Context},
-    stream::SocketInfo,
 };
 
 #[derive(Debug, Clone)]
@@ -59,7 +59,7 @@ impl<State, Body> crate::service::Matcher<State, Request<Body>> for PortMatcher 
 
 impl<State, Socket> crate::service::Matcher<State, Socket> for PortMatcher
 where
-    Socket: crate::stream::Socket,
+    Socket: crate::net::stream::Socket,
 {
     fn matches(
         &self,
@@ -124,7 +124,7 @@ mod test {
             peer_addr: Option<SocketAddr>,
         }
 
-        impl crate::stream::Socket for FakeSocket {
+        impl crate::net::stream::Socket for FakeSocket {
             fn local_addr(&self) -> std::io::Result<SocketAddr> {
                 match &self.local_addr {
                     Some(addr) => Ok(*addr),
