+++ src/net/stream/matcher/socket.rs
@@ -1,8 +1,8 @@
 use http::Request;
 
 use crate::{
+    net::stream::SocketInfo,
     service::{context::Extensions, Context},
-    stream::SocketInfo,
 };
 use std::net::SocketAddr;
 
@@ -54,7 +54,7 @@ impl<State, Body> crate::service::Matcher<State, Request<Body>> for SocketAddres
 
 impl<State, Socket> crate::service::Matcher<State, Socket> for SocketAddressMatcher
 where
-    Socket: crate::stream::Socket,
+    Socket: crate::net::stream::Socket,
 {
     fn matches(
         &self,
@@ -118,7 +118,7 @@ mod test {
             peer_addr: Option<SocketAddr>,
         }
 
-        impl crate::stream::Socket for FakeSocket {
+        impl crate::net::stream::Socket for FakeSocket {
             fn local_addr(&self) -> std::io::Result<SocketAddr> {
                 match &self.local_addr {
                     Some(addr) => Ok(*addr),
