+++ src/net/stream/read.rs
