+++ src/net/stream/service/mod.rs
@@ -1,4 +1,4 @@
-//! Rama services that operate directly on [`crate::stream::Stream`] types.
+//! Rama services that operate directly on [`crate::net::stream::Stream`] types.
 //!
 //! Examples are services that can operate directly on a `TCP`, `TLS` or `UDP` stream.
