+++ src/proxy/http/client/layer.rs
@@ -4,9 +4,9 @@ use crate::http::headers::HeaderMapExt;
 use crate::http::headers::{Authorization, ProxyAuthorization};
 use crate::http::Uri;
 use crate::http::{Request, RequestContext};
+use crate::net::stream::Stream;
 use crate::proxy::{ProxyCredentials, ProxySocketAddr};
 use crate::service::{Context, Layer, Service};
-use crate::stream::Stream;
 use crate::tls::HttpsTunnel;
 use std::fmt;
 use std::future::Future;
